+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.7] - 2023-04-19
+### Added
+* `SetNeighbor(int index, Node neighbor)` and `GetNeighbor(int)` and `HasNeighbor(int)`
+### Changed
+* `Node` property names
+* `Node` neighbors to be ordered by using a sparsely packed list
+* `Node`.`AddNeighbor` to handle sparsely packed list data structure
+### Fixed
+* `AStarPathfinding` bug where `IsValidMoveBetweenNodes()` evaluated incorrect nodes
+
 ## [0.0.6] - 2023-04-12
 ### Changed
 * pooling API calls to Kokowolo.Utilities after updating Utilities package version